Rename microprofile-telemetry-metrics to microprofile-telemetry
