+++ microprofile-telemetry/src/test/resources/arquillian-bootable.xml
